// src/ui/models/ThemeSettings.ts
export type ThemeSettings = {
  // Card
  cardColorH: number;
  cardColorS: number;
  cardColorL: number;
  // Accent Color
  accentColorH: number;
  accentColorS: number;
  accentColorL: number;
  // Text
  textColorH: number;
  textColorS: number;
  textColorL: number;

  focusColorA: number;
  dividerColorA: number;
};